[MODIFY]Update SQL table structures and add update_log table
Modified SQL tables to include 'id' as primary key and remove the NOT NULL constraint from several fields. Furthermore, added DROP TABLE IF EXISTS statement at the beginning of each table creation to ensure older versions are removed. Introduced a new 'update_log' table for logging updates status. This is part of an effort to make the database schema more efficient and easier to manage, as well as provide better logging for updates.